Functions

MeshSteward.initbox — Function

initbox(x::AbstractVector{T}) where {T}

Create a bounding box and initialize with a single point.

MeshSteward.updatebox! — Function

updatebox!(box::AbstractVector{T}, x::AbstractVector{T}) where {T}

Update a box with another location, or create a new box.

If the box does not have the correct dimensions, it is correctly sized.

box = bounding box for 1-D box=[minx,maxx], or for 2-D box=[minx,maxx,miny,maxy], or for 3-D box=[minx,maxx,miny,maxy,minz,maxz] x = vector defining a point. The box is expanded to include the supplied location x.

MeshSteward.boundingbox — Function

boundingbox(x::AbstractArray{T}) where {T}

Compute the bounding box of the points in x.

x = holds points, one per row.

Returns box = bounding box for 1-D box=[minx,maxx], or for 2-D box=[minx,maxx,miny,maxy], or for 3-D box=[minx,maxx,miny,maxy,minz,maxz]

MeshSteward.inflatebox! — Function

inflatebox!(box::AbstractVector{T}, inflatevalue::T) where {T}

Inflate the box by the value supplied.

MeshSteward.inbox — Function

inbox(box::AbstractVector{T}, x::AbstractVector{T}) where {T}

Is the given location inside the box?

box = vector entries arranged as minx,maxx,miny,maxy,minz,maxz.

Note: point on the boundary of the box is counted as being inside.

MeshSteward.boxesoverlap — Function

boxesoverlap(box1::AbstractVector{T}, box2::AbstractVector{T}) where {T}

Do the given boxes overlap?

MeshSteward.intersectboxes — Function

intersectboxes(box1::AbstractVector{T}, box2::AbstractVector{T}) where {T}

Compute the intersection of two boxes.

The function returns an empty box (length(b) == 0) if the intersection is empty; otherwise a box is returned.

MeshSteward.eselect — Function

eselect(ir::IncRel; kwargs...)

Select finite elements.

Arguments

ir = incidence relation representing finite element set (d, 0).

kwargs = keyword arguments to specify the selection criteria

Selection criteria

facing

Select all "boundary" elements that "face" a certain direction.

exteriorbfl = selectelem(fens, bdryfes, facing=true, direction=[1.0, 1.0, 0.0]);

or

exteriorbfl = selectelem(fens, bdryfes, facing=true, direction=dout, dotmin = 0.99);

where

function dout(xyz)
    return xyz/norm(xyz)
end

and xyz is the location of the centroid of a boundary element. Here the finite element is considered "facing" in the given direction if the dot product of its normal and the direction vector is greater than dotmin. The default value for dotmin is 0.01 (this corresponds to almost 90 degrees between the normal to the finite element and the given direction).

This selection method makes sense only for elements that are surface-like (i. e. for boundary mmeshes).

label

Select elements based on their label.

rl1 = selectelem(fens, fes, label=1)

box, distance

Select elements based on some criteria that their nodes satisfy. See the function selectnode().

Example: Select all elements whose nodes are closer than R+inflate from the point from.

linner = selectelem(fens, bfes, distance = R, from = [0.0 0.0 0.0],
  inflate = tolerance)

Example:

exteriorbfl = selectelem(fens, bdryfes,
   box=[1.0, 1.0, 0.0, pi/2, 0.0, Thickness], inflate=tolerance);

flood

Select all FEs connected together, starting from a given node.

Example: Select all FEs connected together (Starting from node 13):

l = selectelem(fens, fes, flood = true, startnode = 13)

Optional keyword arguments

Should we consider the element only if all its nodes are in?

allin = Boolean: if true, then all nodes of an element must satisfy the criterion; otherwise one is enough.

Output

felist = list of finite elements (shapes) from the from the collection on the left of the incidence relation that satisfy the criteria

MeshSteward.vselect — Function

vselect(v::FO; kwargs...) where {FO}

Select locations (vertices) based on some criterion.

FO is the type of a function or an callable object that returns the coordinates of a vertex given its number.

Selection criteria

box

nLx = vselect(geom.co, box = [0.0 Lx  0.0 0.0 0.0 0.0], inflate = Lx/1.0e5)

The keyword 'inflate' may be used to increase or decrease the extent of the box (or the distance) to make sure some nodes which would be on the boundary are either excluded or included.

Returns The list of vertices that match the search criterion.

vselect(m::Mesh; kwargs...)

Select vertices. Return as an incidence relation.

Refer to MeshFinder vselect.

MeshSteward.import_MESH — Function

import_MESH(meshfile)

Import vertices and shapes in the MESH format.

Output

Data dictionary, with keys

"vertices" (vertices),

"shapes" (array of shape collections).

MeshSteward.import_NASTRAN — Function

import_NASTRAN(filename)

Import tetrahedral (4- and 10-node) NASTRAN mesh (.nas file).

Limitations:

only the GRID and CTETRA sections are read.

Only 4-node and 10-node tetrahedra are handled.

The file should be free-form (data separated by commas).

Some fixed-format files can also be processed (large-field, but not small-field).

Output

Data dictionary, with keys

"vertices" (vertices),

"shapes" (array of shape collections).

MeshSteward.import_ABAQUS — Function

import_ABAQUS(filename)

Import tetrahedral (4- and 10-node) or hexahedral (8- and 20-node) Abaqus mesh (.inp file).

Limitations:

Only the *NODE and *ELEMENT sections are read

Only 4-node and 10-node tetrahedra, 8-node or 20-node hexahedra, 3-node triangles are handled.

Output

Data dictionary, with keys

"fens" (finite element nodes),

"fesets" (array of finite element sets).

MeshSteward.vtkwrite — Function

vtkwrite(filename, connectivity)

Write VTK file with the mesh.

connectivity = incidence relation of the type d -> 0. It must have a "geom" attribute for access to the locations of all the vertices that the connectivity incidence relation references.

data = array of string keys. Names of attributes of either the left or the right shape collection of the connectivity incidence relation.

MeshSteward.export_MESH — Function

export_MESH(meshfile, mesh)

Import vertices and shapes in the MESH format.

Output

Data dictionary, with keys

"vertices" (vertices),

"shapes" (array of shape collections).

MeshSteward.load — Function

load(m::Mesh, filename::String)

Load a mesh (incidence relation) from a MESH file.

Note

No check is performed that the loaded incidence relation is compatible with the existing incidence relations in the mesh.

MeshSteward.save — Function

save(m::Mesh, filename::String)

Save a mesh base incidence relation to a MESH file.

MeshSteward.increl — Function

increl(m::Mesh, irc::Tuple{Int64, Int64})

Retrieve the named incidence relation.

MeshSteward.insert! — Function

insert!(m::Mesh, increl::IncRel)

Insert the incidence relation. The code of the incidence relation is the key under which this relation is stored in the mesh.

MeshSteward.basecode — Function

basecode(m::Mesh)

Compute the code of the base relation.

The base incidence relation is (d, 0) that represents the elements of the interior of the domain.

MeshSteward.nspacedims — Function

nspacedims(m::Mesh)

Furnish the dimension of the space in which the mesh lives.

Base.summary — Function

Base.summary(ir::IncRel)

Form a brief summary of the incidence relation.

Base.summary(sc::S) where {S<:ShapeColl}

Form a brief summary of the shape collection.

Base.summary(m::Mesh)

Form a brief summary of the mesh.

Missing docstring.

Missing docstring for vselect. Check Documenter's build log for details.

MeshSteward.boundary — Function

boundary(m::Mesh)

Compute the boundary of the mesh.

MeshSteward.vertices — Function

vertices(m::Mesh)

Compute the (0, 0) incidence relation for the vertices of the base incidence relation.

Missing docstring.

Missing docstring for submesh. Check Documenter's build log for details.

Missing docstring.

Missing docstring for label. Check Documenter's build log for details.

Index

Base.summary

MeshSteward.basecode

MeshSteward.boundary

MeshSteward.boundingbox

MeshSteward.boxesoverlap

MeshSteward.eselect

MeshSteward.export_MESH

MeshSteward.import_ABAQUS

MeshSteward.import_MESH

MeshSteward.import_NASTRAN

MeshSteward.inbox

MeshSteward.increl

MeshSteward.inflatebox!

MeshSteward.initbox

MeshSteward.insert!

MeshSteward.intersectboxes

MeshSteward.load

MeshSteward.nspacedims

MeshSteward.save

MeshSteward.updatebox!

MeshSteward.vertices

MeshSteward.vselect

MeshSteward.vtkwrite